[refactor/#7]: service 구현체를 사용하도록 변경
